migrate to BigDecimal format

+++ src/main/java/com/airwallex/assignment/operations/Operation.java
@@ -4,5 +4,5 @@ import java.util.Stack;
 
 @FunctionalInterface
 public interface Operation {
-    void execute(Stack<Double> stack, Stack<Stack<Double>> history);
+    void execute(Stack<BigDecimal> stack, Stack<Stack<BigDecimal>> history);
 }
